moilapp-develop_anto/requirements.txt
PyQt6~=6.3.1
opencv-python~=4.6.0.66
numpy~=1.22.4
dropbox
pyexiv2
GitPython
netifaces
PyYAML~=6.0